+++ DependencyInjection/CompilerPass/ImageFormatCompilerPass.php
@@ -11,8 +11,11 @@
 
 namespace Sulu\Bundle\ThemeBundle\DependencyInjection\CompilerPass;
 
-use Sulu\Bundle\MediaBundle\Media\FormatLoader\XmlFormatLoader;
+use Sulu\Bundle\MediaBundle\Media\FormatLoader\XmlFormatLoader10;
+use Sulu\Bundle\MediaBundle\Media\FormatLoader\XmlFormatLoader11;
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\Config\Loader\DelegatingLoader;
+use Symfony\Component\Config\Loader\LoaderResolver;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
@@ -46,11 +49,11 @@ class ImageFormatCompilerPass implements CompilerPassInterface
     }
 
     /**
-     * @param array $defaultOptions
+     * @param array $globalOptions
      *
      * @return array
      */
-    protected function loadThemeFormats($defaultOptions)
+    private function loadThemeFormats($globalOptions)
     {
         $activeFormats = [];
         $activeTheme = $this->container->get('liip_theme.active_theme');
@@ -73,7 +76,7 @@ class ImageFormatCompilerPass implements CompilerPassInterface
                 );
 
                 if (file_exists($fullPath)) {
-                    $this->setFormatsFromFile($fullPath, $activeFormats, $defaultOptions);
+                    $this->setFormatsFromFile($fullPath, $activeFormats, $globalOptions);
                 }
             }
         }
@@ -84,21 +87,26 @@ class ImageFormatCompilerPass implements CompilerPassInterface
     /**
      * @param $fullPath
      * @param $activeFormats
-     * @param $defaultOptions
+     * @param $globalOptions
      */
-    protected function setFormatsFromFile($fullPath, &$activeFormats, $defaultOptions)
+    private function setFormatsFromFile($fullPath, &$activeFormats, $globalOptions)
     {
         $folder = dirname($fullPath);
         $fileName = basename($fullPath);
 
         $locator = new FileLocator($folder);
-        $loader = new XmlFormatLoader($locator);
-        $loader->setDefaultOptions($defaultOptions);
+
+        $xmlLoader10 = new XmlFormatLoader10($locator);
+        $xmlLoader11 = new XmlFormatLoader11($locator);
+        $xmlLoader10->setGlobalOptions($globalOptions);
+        $xmlLoader11->setGlobalOptions($globalOptions);
+
+        $resolver = new LoaderResolver([$xmlLoader10, $xmlLoader11]);
+        $loader = new DelegatingLoader($resolver);
+
         $themeFormats = $loader->load($fileName);
         foreach ($themeFormats as $format) {
-            if (isset($format['name']) && !array_key_exists($format['name'], $activeFormats)) {
-                $activeFormats[$format['name']] = $format;
-            }
+            $activeFormats[$format['key']] = $format;
         }
     }
 }