+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 CHANGELOG for Sulu
 ==================
 
+* dev-develop
+    * ENHANCEMENT #2 Refactored ImageFormatCompilerPass to meet the requirements of the new image formats definition
+
 * 1.0.0 (2016-07-22)
-    * FEATURE #1 Introduced liip/theme-bundle to enable theming
+    * FEATURE     #1 Introduced liip/theme-bundle to enable theming